Data type conversion
Some data type where wrong. Signature has beed redefined for the use of
size_t, ssize_t, etc...

+++ knapsackkbest/src/kbest.h
@@ -14,7 +14,7 @@
 #include "data.h"
 
 struct kProblem {
-	uint32_t num_var;
+	size_t num_var;
 	uint32_t* weights;
 	uint32_t* values;
 	uint32_t max_weigth;
@@ -24,8 +24,8 @@ typedef struct kProblem* KProblem;
 /*
  * Initialize a knapsack problem.
  */
-void kp_init_kp(KProblem* problem, uint32_t n, uint32_t* weights, uint32_t* values,
-uint32_t maxW);
+void kp_init_kp(KProblem* problem, size_t n, uint32_t* weights, uint32_t*
+    values, uint32_t maxW);
 /*
  * Free a knapsack problem.
  */
@@ -33,7 +33,7 @@ void kp_free_kp(KProblem problem);
 
 struct kSolution {
 	uint32_t tot_value;
-	uint32_t vector_size;
+	size_t vector_size;
 	uint32_t* solution_vector;
 };
 typedef struct kSolution* KSolution;
@@ -41,7 +41,7 @@ typedef struct kSolution* KSolution;
 
 struct kBestSolutions {
 	KProblem problem;
-	uint32_t sol_count;
+	size_t sol_count;
 	KSolution* solutions;
 };
 typedef struct kBestSolutions* KBestSolutions;
@@ -49,7 +49,7 @@ typedef struct kBestSolutions* KBestSolutions;
 /*
  * Initialize a knapsack solution.
  */
-void kp_init_sol(KSolution* solution, uint32_t n);
+void kp_init_sol(KSolution* solution, size_t n);
 /*
  * Free a knapsack solution.
  */
@@ -58,7 +58,7 @@ void kp_free_sol(KSolution solution);
 /*
  * Initialize a knapsack solution.
  */
-void kp_init_kbest_sols(KBestSolutions* solution, KProblem problem, uint32_t sol_count);
+void kp_init_kbest_sols(KBestSolutions* solution, KProblem problem, size_t sol_count);
 /*
  * Free a knapsack solution.
  */
@@ -67,6 +67,6 @@ void kp_free_kbest_sols(KBestSolutions solution);
 /*
  * Solve the knapsack problem, finding best_sol_count solutions.
  */
-void kp_solve(KBestSolutions* dest, KProblem problem, uint32_t best_sol_count);
+void kp_solve(KBestSolutions* dest, KProblem problem, size_t best_sol_count);
 
 #endif /* KBEST_H_ */
